+++ store/thunks/dashboard.ts
@@ -1,8 +1,9 @@
 import { createAsyncThunk } from '@reduxjs/toolkit';
 
 import { db } from '../../db';
+import * as api from 'service/apiService';
 
-import { ThunkApi, RatingCounts, TagHistory } from 'store/types';
+import { ThunkApi, RatingCounts, TagHistory, FoundTags, NotFoundTags } from 'store/types';
 import { Tag, Post } from 'types/gelbooruTypes';
 
 export const fetchDownloadedPostCount = createAsyncThunk<number, void, ThunkApi>(
@@ -22,7 +23,8 @@ export const fetchBlacklistedPostCount = createAsyncThunk<number, void, ThunkApi
 export const fetchFavoritePostCount = createAsyncThunk<number, void, ThunkApi>(
 	'dashboard/fetchFavoritePostCount',
 	async (): Promise<number> => {
-		return db.posts.getFavoriteCount();
+		const postIds = await db.favoritesTree.getlAllPostIds();
+		return postIds.length;
 	}
 );
 
@@ -50,10 +52,36 @@ export const fetchMostSearchedTags = createAsyncThunk<TagHistory[], void, ThunkA
 	}
 );
 
-export const fetchMostFavoritedTags = createAsyncThunk<{ tag: Tag | undefined; count: number }[], number | undefined, ThunkApi>(
+export const fetchMostFavoritedTags = createAsyncThunk<{ tag: Tag; count: number }[], number | undefined, ThunkApi>(
 	'dashboard/fetchMostFavoritedTags',
-	async (limit = 20): Promise<{ tag: Tag | undefined; count: number }[]> => {
-		return db.tags.getMostFavorited(limit);
+	async (limit = 20, thunkApi): Promise<{ tag: Tag; count: number }[]> => {
+		const tags = await db.favoritesTree.getAllFavoriteTagsWithCounts();
+		const sorted = tags.sort((a, b) => b.count - a.count).slice(0, limit < 100 ? limit : 100);
+
+		const notFoundTags: NotFoundTags[] = [];
+		const foundTags: FoundTags[] = [];
+
+		await Promise.all(
+			sorted.map(async (tag) => {
+				const tagFromDb = await db.tags.getTag(tag.tag);
+				if (tagFromDb) {
+					foundTags.push({ tag: tagFromDb, count: tag.count });
+				} else {
+					notFoundTags.push({ tag: tag.tag, count: tag.count });
+				}
+			})
+		);
+
+		const tagsFromApi = await api.getTagsByNames(
+			notFoundTags.map((tag) => tag.tag),
+			thunkApi.getState().settings.apiKey
+		);
+		tagsFromApi.forEach((tag) => {
+			const notFoundTag = notFoundTags.find((nf) => nf.tag === tag.tag);
+			if (!notFoundTag) throw new Error('Error while fetching most favorited tags while joining tags from api with counts.');
+			foundTags.push({ tag, count: notFoundTag.count });
+		});
+		return foundTags.sort((a, b) => b.count - a.count);
 	}
 );